Refactor Next.js components to deep client nesting with granular structure
